Update given phase tests to reflect the latest changes
