+++ game/src/test/java/net/theevilreaper/tamias/game/phase/lobby/LobbyPhaseConditionIntegrationTest.java
@@ -1,19 +1,14 @@
-package net.theevilreaper.tamias.game.phase;
+package net.theevilreaper.tamias.game.phase.lobby;
 
 import net.minestom.server.entity.Player;
 import net.minestom.server.instance.Instance;
 import net.minestom.testing.Env;
 import net.minestom.testing.extension.MicrotusExtension;
-import net.theevilreaper.aves.map.provider.MapProvider;
-import net.theevilreaper.tamias.game.map.GameMapProvider;
 import org.jetbrains.annotations.NotNull;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
-import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.ThreadLocalRandom;
@@ -22,23 +17,7 @@ import static org.junit.jupiter.api.Assertions.*;
 
 @Disabled("Needs a refactor")
 @ExtendWith(MicrotusExtension.class)
-class LobbyPhaseConditionIntegrationTest {
-
-    private static final int LOBBY_PHASE_TIME = 60;
-    private static LobbyPhase lobbyPhase;
-
-    @BeforeAll
-    static void setup() {
-        MapProvider mapProvider = new GameMapProvider(Paths.get(""));
-        lobbyPhase = new LobbyPhase(mapProvider, value -> {}, () -> {}, 1, 2, LOBBY_PHASE_TIME);
-    }
-
-    @AfterEach
-    void cleanup() {
-        lobbyPhase.setPaused(true);
-        lobbyPhase.setCurrentTicks(LOBBY_PHASE_TIME);
-        lobbyPhase.setForceStarted(false);
-    }
+class LobbyPhaseConditionIntegrationTest extends LobbyPhaseTestBase {
 
     @Test
     void testLobbyPhasePlayerUpdate(@NotNull Env env) {
