+++ game/src/test/java/net/theevilreaper/tamias/game/phase/lobby/LobbyPhaseIntegrationTest.java
@@ -1,16 +1,10 @@
-package net.theevilreaper.tamias.game.phase;
+package net.theevilreaper.tamias.game.phase.lobby;
 
 import net.minestom.testing.extension.MicrotusExtension;
-import net.theevilreaper.aves.map.provider.MapProvider;
 import net.theevilreaper.tamias.common.config.GameConfig;
-import net.theevilreaper.tamias.game.map.GameMapProvider;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-
-import java.nio.file.Paths;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -19,25 +13,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @Disabled("Needs a refactor")
 @ExtendWith(MicrotusExtension.class)
-class LobbyPhaseIntegrationTest {
-
-    private static final int LOBBY_PHASE_TIME = 60;
-    private static LobbyPhase lobbyPhase;
-
-    @BeforeAll
-    static void setup() {
-        MapProvider mapProvider = new GameMapProvider(Paths.get(""));
-        lobbyPhase = new LobbyPhase(mapProvider, value -> {
-        }, () -> {
-        }, 1, 2, LOBBY_PHASE_TIME);
-    }
-
-    @AfterEach
-    void cleanup() {
-        lobbyPhase.setPaused(true);
-        lobbyPhase.setCurrentTicks(LOBBY_PHASE_TIME);
-        lobbyPhase.setForceStarted(false);
-    }
+class LobbyPhaseIntegrationTest extends LobbyPhaseTestBase {
 
     @Test
     void testForceStart() {
